Add execute declare.
