+++ crates/blockifier/src/transaction.rs
@@ -1,8 +1,7 @@
 pub mod account_transaction;
 pub mod constants;
-pub mod deploy_account_transaction;
 pub mod errors;
 pub mod execute_transaction;
-pub mod invoke_transaction;
 pub mod objects;
 pub mod transaction_utils;
+pub mod transactions;
